urlshortener/requirements.txt
django_extensions==3.1.5
django>=2.2
djangorestframework==3.13.1
ipdb==0.13.9